+++ test-root/models.py
@@ -30,7 +30,9 @@ class ExampleModel(Model):
     """
 
     foo = StringType(
-        required=True, metadata=dict(custom_value=True), choices=("fizz", "buzz")
+        required=True,
+        metadata=dict(custom_value=True, another_value=dict(one="1", two="2", apple="3")),
+        choices=("fizz", "buzz"),
     )
 
     #: This is bar's docstring