+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Ensure we automatically document ModelType fields nested in ListType and DictType fields
 * Only document each model class once
+* All field and metadata output is now sorted